less from book

+++ interuct.py
@@ -2,7 +2,7 @@ while True:
     reply = input('Введите текст:')
     if reply == 'стоп': break
     try:
-        print(int(reply)**2)
+        print(float(reply)**2)
     except:
         print('bad'*8)
 print('bye')
